📝 Pugrade documentation
