+++ compiler/IR/Instr/IRInstrArithmeticOp.h
@@ -1,92 +1,111 @@
 #pragma once
 
+#include "BaseIRInstr.h"
 #include "IRInstrBinaryOp.h"
-#include <ostream>
 
 /**
- * @brief Represents a comparison operation instruction in the intermediate representation.
+ * @brief Represents an arithmetic operation instruction in the intermediate representation.
  *
- * This class handles the generation of intermediate representation instructions for comparison operations,
- * such as equality, inequality, greater than, greater than or equal to, less than, and less than or equal to.
+ * This class handles the generation of intermediate representation instructions for arithmetic operations,
+ * such as addition, subtraction, multiplication, division, modulo, and bitwise operations. It extends the
+ * `IRInstrBinaryOp` class and provides specialized methods for handling these operations.
  */
-class IRInstrComp : public IRInstrBinaryOp
+class IRInstrArithmeticOp : public IRInstrBinaryOp
 {
 public:
     /**
-     * @brief Constructs a comparison operation instruction.
+     * @brief Constructs an arithmetic operation instruction.
      *
-     * Initializes the instruction with a basic block, two operands, and the comparison operation.
+     * Initializes the instruction with a basic block, two operands, and the arithmetic operation.
      *
      * @param bb_ The basic block to which the instruction belongs.
-     * @param firstOp The first operand of the comparison operation.
-     * @param secondOp The second operand of the comparison operation.
-     * @param op The comparison operation (e.g., '==', '!=', '>', '<', '>=', '<=').
+     * @param firstOp The first operand of the arithmetic operation.
+     * @param secondOp The second operand of the arithmetic operation.
+     * @param op The arithmetic operation (e.g., '+', '-', '*', '/', '%').
      */
-    IRInstrComp(BasicBlock *bb_, string firstOp, string secondOp, string op)
+    IRInstrArithmeticOp(BasicBlock *bb_, string firstOp, string secondOp, string op)
         : IRInstrBinaryOp(bb_, firstOp, secondOp, op) {}
 
     /**
-     * @brief Generates the assembly code for this comparison operation instruction.
+     * @brief Generates the assembly code for this arithmetic operation instruction.
      *
-     * This method generates the appropriate assembly code based on the specific comparison operation
-     * (e.g., equality, inequality, greater than, greater than or equal to, less than, or less than or equal to).
+     * This method generates the appropriate assembly code based on the specific arithmetic operation
+     * (e.g., addition, subtraction, multiplication, division, modulo, or bitwise operations).
      *
      * @param o The output stream where the generated assembly code will be written.
      */
-    virtual void gen_asm(ostream &o) override;
+    virtual void gen_asm(ostream &o);
 
 private:
     /**
-     * @brief Handles the equality comparison operation (`==`).
+     * @brief Handles the addition operation.
      *
-     * This method generates assembly code for the equality comparison operation.
+     * This method generates assembly code for the addition operation (`+`).
      *
      * @param o The output stream where the generated assembly code will be written.
      */
-    void handleCompEq(ostream &o);
+    void handleAddition(ostream &o);
 
     /**
-     * @brief Handles the inequality comparison operation (`!=`).
+     * @brief Handles the subtraction operation.
      *
-     * This method generates assembly code for the inequality comparison operation.
+     * This method generates assembly code for the subtraction operation (`-`).
      *
      * @param o The output stream where the generated assembly code will be written.
      */
-    void handleCompNotEq(ostream &o);
+    void handleSubstraction(ostream &o);
 
     /**
-     * @brief Handles the greater-than comparison operation (`>`).
+     * @brief Handles the multiplication operation.
      *
-     * This method generates assembly code for the greater-than comparison operation.
+     * This method generates assembly code for the multiplication operation (`*`).
      *
      * @param o The output stream where the generated assembly code will be written.
      */
-    void handleGreater(ostream &o);
+    void handleMult(ostream &o);
 
     /**
-     * @brief Handles the greater-than or equal-to comparison operation (`>=`).
+     * @brief Handles the division operation.
      *
-     * This method generates assembly code for the greater-than or equal-to comparison operation.
+     * This method generates assembly code for the division operation (`/`).
      *
      * @param o The output stream where the generated assembly code will be written.
      */
-    void handleGreaterOrEqual(ostream &o);
+    void handleDiv(ostream &o);
 
     /**
-     * @brief Handles the less-than comparison operation (`<`).
+     * @brief Handles the modulo operation.
      *
-     * This method generates assembly code for the less-than comparison operation.
+     * This method generates assembly code for the modulo operation (`%`).
      *
      * @param o The output stream where the generated assembly code will be written.
      */
-    void handleLower(ostream &o);
+    void handleModulo(ostream &o);
 
     /**
-     * @brief Handles the less-than or equal-to comparison operation (`<=`).
+     * @brief Handles the bitwise AND operation.
      *
-     * This method generates assembly code for the less-than or equal-to comparison operation.
+     * This method generates assembly code for the bitwise AND operation (`&`).
      *
      * @param o The output stream where the generated assembly code will be written.
      */
-    void handleLowerOrEqual(ostream &o);
+    void handleBitwiseAnd(ostream &o);
+
+    /**
+     * @brief Handles the bitwise OR operation.
+     *
+     * This method generates assembly code for the bitwise OR operation (`|`).
+     *
+     * @param o The output stream where the generated assembly code will be written.
+     */
+    void handleBitwiseOr(ostream &o);
+
+    /**
+     * @brief Handles the bitwise XOR operation.
+     *
+     * This method generates assembly code for the bitwise XOR operation (`^`).
+     *
+     * @param o The output stream where the generated assembly code will be written.
+     */
+    void handleBitwiseXor(ostream &o);
 };
